requirements.txt
arcticdb>=5.0.0
pandas>=2.0.0
pyarrow>=14.0.0
pyyaml>=6.0
requests>=2.31.0
openpyxl>=3.1.2
python-docx>=1.1.0
jinja2>=3.1.0
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.13.0